Update RPS.js
computerPlay() randomly generates a choice for the computer.

playRound() checks the user's input against the computer's choice and increments scores accordingly. Additionally, it returns a message denoting if the user won or lost

game() takes a number (num) and repeats playRound num times, printing the game results to the console. Once the loop is finished, the playerScore and computerScore are checked against one another and a message detailing how many games were won and tied is printed to the console.